+++ client/game/Research/Queue/Queue.vue
@@ -1,13 +1,15 @@
 <script>
 /***********************************************************************************************************************
- * shows the research qeue
+ * shows the research queue
  **********************************************************************************************************************/
 import draggable from "vuedraggable";
-import QeueItem from "./QeueItem.vue";
+import QueueItem from "./QueueItem.vue";
+import AbortResearch from "./AbortResearch.vue";
 export default {
     components: {
         draggable,
-        QeueItem
+        QueueItem,
+        AbortResearch
     },
     computed: {
         researches: {
@@ -21,17 +23,21 @@ export default {
 
 <template>
     <section>
-        <h2>Research Qeue</h2>
+        <h2>{{ $t("queue.label") }}</h2>
         <draggable
             v-if="researches.length"
-            v-model="researches">
-            <qeue-item
+            v-model="researches"
+            :options="{disabled: isChangingOrder}">
+            <queue-item
                 v-for="research in researches"
                 :key="research.id"
                 :research-id="research.id" />
         </draggable>
+        <abort-research
+            v-for="research in researches"
+            :key="research.id"
+            :research-id="research.id" />
     </section>
-
 </template>
 
 <style lang="scss" scoped>