+++ client/game/Research/TechLevels/LevelOverview.spec.js
@@ -4,7 +4,7 @@
 import Vuex from "vuex";
 import VueI18n from "vue-i18n";
 import {shallow, createLocalVue} from "@vue/test-utils";
-import TechLevel from "./TechLevel.vue";
+import LevelOverview from "./LevelOverview.vue";
 
 /*
  * mock data
@@ -16,8 +16,11 @@ const i18n = new VueI18n({
     locale: "de",
     messages: {
         de: {
-            techLevels: {shields: "Schutzschilde", laser: "Laser", armour: "Panzerung", missile: "Raketen"},
-            research: {start: "TL{tl} Erforschen"}
+            techLevels: {
+                shields: "Shields",
+                laser: "Laser",
+                status: {researched: "researched", queued: "in queue", unresearched: "unresearched"}
+            }
         }
     }
 });
@@ -25,12 +28,7 @@ const store = new Vuex.Store({
     state: {},
     getters: {
         playerResearches() {
-            return [
-                {area: "laser", order: 0, newLevel: 1},
-                {area: "armour", order: 1, newLevel: 1},
-                {area: "shields", order: 2, newLevel: 5},
-                {area: "shields", order: 3, newLevel: 6}
-            ];
+            return [{area: "shields", order: 0, newLevel: 5}, {area: "shields", order: 1, newLevel: 6}];
         }
     }
 });
@@ -38,11 +36,11 @@ const store = new Vuex.Store({
 /*
  * test suite
  */
-describe("TechLevel.vue", () => {
+describe("LevelOverview.vue", () => {
     let cmp;
 
     beforeEach(() => {
-        cmp = shallow(TechLevel, {
+        cmp = shallow(LevelOverview, {
             localVue,
             i18n,
             store,
@@ -63,13 +61,15 @@ describe("TechLevel.vue", () => {
 
     it("calculates the correct next level", () => {
         expect(cmp.vm.nextLevel).toBe(7);
-        cmp.setProps({tlType: "laser", level: 0});
+        cmp.setProps({tlType: "laser", level: 1});
         expect(cmp.vm.nextLevel).toBe(2);
-        cmp.setProps({tlType: "missile", level: 0});
-        expect(cmp.vm.nextLevel).toBe(1);
     });
 
-    it("shows the correct headline for the tech level type", () => {
-        expect(cmp.find("h4").text()).toBe("Schutzschilde");
+    it("shows the job with order=0 as 'researching'", () => {
+        expect(cmp.vm.getResearchedClass(5)).toBe("unresearched queued researching");
+    });
+
+    it("shows the correct number of unresearched items", () => {
+        expect(cmp.findAll(".unresearched").length).toBe(6);
     });
 });