+++ client/game/Research/store/actions.js
@@ -37,7 +37,7 @@ const ACTIONS = {
     /*
      * fetch game data from api via XHR
      * @param {Object} ctx - Vuex $store context
-     * @param {Object} payload - Array of full research objects
+     * @param {Array} payload - Array of full research objects
      */
     CHANGE_RESEARCH_ORDER: function(ctx, payload) {
         let pos = 0;
@@ -48,20 +48,51 @@ const ACTIONS = {
         });
         DEBUG && console.log(`requesting change research order ${payload.map(res => `${res.area} TL${res.newLevel}`)}`);
         ctx.commit("IS_CHANGING_ORDER", true);
-        ctx.commit("SET_RESEARCHES", data); // we pre-emptively change the state to appear `snappier`
+        // we pre-emptively change the state to be `snappier`. we roll back the changes in case of error
+        ctx.commit("SET_RESEARCHES", data);
         axios
             .post("/api/game/research/order", data)
             .then(response => {
-                if (response.status === 200 && response.data) {
+                if (response.data && !response.data.error) {
                     ctx.commit("SET_RESEARCHES", response.data);
+                }
+                if (response.data && response.data.error) { // server has error message ?
+                    this._vm.$snotify.error(response.data.error);
+                    ctx.commit("SET_RESEARCHES", response.data.researches);
                 }
                 ctx.commit("IS_CHANGING_ORDER", false);
             })
             .catch(error => {
                 console.error(error);
                 ctx.commit("IS_CHANGING_ORDER", false);
+                ctx.commit("ROLLBACK_RESEARCHES");
             });
-    }
+    },
+
+    /*
+     * request "delete research job"
+     * @param {Object} ctx - Vuex $store context
+     * @param {String} payload - ID of research job
+     */
+    DELETE_RESEARCH_JOB: function(ctx, payload) {
+        DEBUG && console.log(`requesting delete research job #${payload}`);
+        ctx.commit("DELETING_RESEARCH_JOB", {id: payload, deleting: true});
+        axios
+            .post("/api/game/research/delete", { id: payload })
+            .then(response => {
+                if (response.data && !response.data.error) {
+                    ctx.commit("SET_RESEARCHES", response.data);
+                }
+                if (response.data && response.data.error) { // server has error message ?
+                    this._vm.$snotify.error(response.data.error);
+                }
+                ctx.commit("DELETING_RESEARCH_JOB", {id: payload, deleting: false});
+            })
+            .catch(error => {
+                console.error(error);
+                ctx.commit("DELETING_RESEARCH_JOB", {id: payload, deleting: false});
+            });
+    },
 };
 
 export default ACTIONS;